tests/the71/requirements.txt
# GhostMesh Testing Framework Requirements
# For THE-71 Comprehensive Testing and Validation

# Core dependencies
paho-mqtt>=1.6.0
numpy>=1.21.0
psutil>=5.8.0
requests>=2.25.0

# Optional dependencies for enhanced testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
